[#713] Refactor Step 5 scenario modeling into StandardScenarioModeling component
