+++ frontend/src/pages/cases/components/index.js
@@ -22,5 +22,6 @@ export { default as TwoDriverHeatmap } from "./TwoDriverHeatmap";
 export { default as TwoBinningDriverForm } from "./TwoBinningDriverForm";
 export { default as ExploreChangeToCloseTheGap } from "./ExploreChangeToCloseTheGap";
 export { default as ThreeDriverCalculator } from "./ThreeDriverCalculator";
+export { default as StandardScenarioModeling } from "./StandardScenarioModeling";
 
 export const MAX_SEGMENT = 5;
